requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
